+++ livechat/agent/web/api/v37.py
@@ -483,6 +483,31 @@ class AgentWebV37(HttpClient):
                                  json=payload,
                                  headers=headers)
 
+    def request_thread_summary(self,
+                               chat_id: str = None,
+                               thread_id: str = None,
+                               payload: dict = None,
+                               headers: dict = None) -> httpx.Response:
+        ''' Requests a thread summary.
+
+            Args:
+                chat_id (str): ID of the chat to request thread summary for.
+                thread_id (str): ID of the thread to request summary for.
+                payload (dict): Custom payload to be used as request's data.
+                                It overrides all other parameters provided for the method.
+                headers (dict): Custom headers to be used with session headers.
+                                They will be merged with session-level values that are set,
+                                however, these method-level parameters will not be persisted across requests.
+
+            Returns:
+                httpx.Response: The Response object from `httpx` library,
+                                which contains a server's response to an HTTP request. '''
+        if payload is None:
+            payload = prepare_payload(locals())
+        return self.session.post(f'{self.api_url}/request_thread_summary',
+                                 json=payload,
+                                 headers=headers)
+
 # Properties
 
     def update_chat_properties(self,